[codeshuffle] - moved XbmcContext.h/.cpp
