+++ xbmc/Makefile.in
@@ -35,7 +35,6 @@ SRCS=Application.cpp \
      XBApplicationEx.cpp \
      XBDateTime.cpp \
      xbmc.cpp \
-     XbmcContext.cpp \
 
 LIB=xbmc.a
