+++ xbmc/platform/XbmcContext.cpp
