+++ xbmc/platform/XbmcContext.h
